refactor: 데이터 조회를 API 라우트로 분리 및 fetch 방식으로 전환
